Fixed .1 error, 100.1 error, fixed address, changed island
Added isInterger to must be number check to prevent .1 being accepted .

Changed parseInt to parseFloat on box range check to prevent 100.1 being accepted.

Changed parseInt+isInterger not number detection to isNaN to allow numbers and letters to be entered. Just not numbers only.

Changed the error message to reflect this.

Removed isnumber check from island code as island as shipping location could have numbers in the title.

Numbers don't need to be checked for as any shipping location with random numbers wont match location on list.